+++ ripozo/__init__.py
@@ -9,3 +9,4 @@ from ripozo.viewsets import fields, restmixins
 from ripozo.viewsets.relationships.list_relationship import ListRelationship
 from ripozo.viewsets.relationships.relationship import Relationship
 from ripozo.viewsets.resource_base import ResourceBase
+from ripozo.viewsets.request import RequestContainer